next release (#188)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @darkmagic/react@0.36.0

### Minor Changes

-
[`b018100`](https://github.com/magicbell/darkmagic/commit/b0181007e79c51fa6c37c60156d672f2ffb250b4)
Thanks [@smeijer](https://github.com/smeijer)! - Improves the checkbox
component with the following changes:

- Checkbox is now wrapped in a Label, so that clicking gaps still
toggles it
- Added the onClickLabel event, to distinguish clicking the label from
the checkbox
- Added the onMouseEnter and onMouseLeave events, for custom hover
actions

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>